some styles fixes

+++ server/services/spaceService.js
@@ -1,23 +1,88 @@
 const SpaceRepository = require('../repositories/SpaceRepository')
 
 module.exports = {
-  getAll: () => {
-    return SpaceRepository.getAll()
+  findAll: (req, res) => {
+    SpaceRepository.getAll()
+      .then(data => res.json(data))
+      .catch((err) => {
+        console.log(err)
+        res.status(400)
+        res.end()
+      })
   },
 
-  get: (id) => {
-    return SpaceRepository.getById(id)
+  findOne: (req, res) => {
+    const id = req.params.id
+
+    if (id.length === 0) {
+      res.status(400)
+      return res.end('Invalid id')
+    }
+
+    SpaceRepository.getById(id)
+      .then(data => res.json(data))
+      .catch((err) => {
+        console.log(err)
+        res.status(400)
+        res.end()
+      })
   },
 
-  create: (data) => {
-    return SpaceRepository.create(data)
+  add: (req, res) => {
+    if (typeof req.body !== 'object') {
+      res.status(400)
+
+      return res.end('Invalid data')
+    }
+
+    if (!req.body.name || !req.body.ownerId || !req.body.key) {
+      res.status(400)
+
+      return res.end('Missing required param')
+    }
+
+    SpaceRepository.create(req.body)
+      .then(data => res.json(data))
+      .catch((err) => {
+        console.log(err)
+        res.status(400)
+        res.end()
+      })
   },
 
-  update: (id, data) => {
-    return SpaceRepository.update(id, data)
+  findOneAndUpdate: (req, res) => {
+    const id = req.params.id
+
+    if (id.length === 0) {
+      res.status(400)
+
+      return res.end('Invalid id')
+    }
+
+    SpaceRepository.update(id, req.body)
+      .then(data => res.json(data))
+      .catch((err) => {
+        console.log(err)
+        res.status(400)
+        res.end()
+      })
   },
 
-  delete: (id) => {
-    return SpaceRepository.delete(id)
+  findOneAndDelete: (req, res) => {
+    const id = req.params.id
+
+    if (id.length === 0) {
+      res.status(400)
+
+      return res.end('Invalid id')
+    }
+
+    SpaceRepository.delete(id)
+      .then(data => res.json(data))
+      .catch((err) => {
+        console.log(err)
+        res.status(400)
+        res.end()
+      })
   }
 }
